+++ doc/reference.rst
@@ -22,6 +22,7 @@ every module.
    gen_gallery
    backreferences
    gen_rst
+   scrapers
    py_source_parser
    docs_resolv
    notebook
