+++ sphinx_gallery/tests/test_gen_rst.py
@@ -16,16 +16,10 @@ import zipfile
 
 import pytest
 
-import numpy as np
-from PIL import Image
-
 import sphinx_gallery.gen_rst as sg
 from sphinx_gallery import downloads
 from sphinx_gallery.gen_gallery import generate_dir_rst, _complete_gallery_conf
 from sphinx_gallery.utils import _TempDir
-
-# Need to import gen_rst before matplotlib.pyplot to set backend to 'Agg'
-import matplotlib.pyplot as plt
 
 CONTENT = [
     '"""',
@@ -277,101 +271,6 @@ def test_thumbnail_number(test_str):
     assert file_conf == {'thumbnail_number': 2}
 
 
-def test_save_matplotlib_figures(gallery_conf):
-    """Test matplotlib figure save"""
-    plt.plot(1, 1)
-    fname_template = os.path.join(gallery_conf['gallery_dir'], 'image{0}.png')
-    image_path_iterator = sg.ImagePathIterator(fname_template)
-    block = ('',) * 3
-    block_vars = dict(image_path_iterator=image_path_iterator)
-    image_rst = sg.save_figures(block, block_vars, gallery_conf)
-    assert len(image_path_iterator) == 1
-    assert '/image1.png' in image_rst
-
-    # Test capturing 2 images with shifted start number
-    image_path_iterator.next()
-    image_path_iterator.next()
-    plt.plot(1, 1)
-    plt.figure()
-    plt.plot(1, 1)
-    image_rst = sg.save_figures(block, block_vars, gallery_conf)
-    assert len(image_path_iterator) == 5
-    assert '/image4.png' in image_rst
-    assert '/image5.png' in image_rst
-
-
-def test_save_mayavi_figures(gallery_conf):
-    """Test file naming when saving figures. Requires mayavi."""
-    try:
-        from mayavi import mlab
-    except ImportError:
-        raise pytest.skip('Mayavi not installed')
-    mlab.options.offscreen = True
-
-    gallery_conf.update(
-        image_scrapers=(sg.matplotlib_scraper, sg.mayavi_scraper))
-    fname_template = os.path.join(gallery_conf['gallery_dir'], 'image{0}.png')
-    image_path_iterator = sg.ImagePathIterator(fname_template)
-    block = ('',) * 3
-    block_vars = dict(image_path_iterator=image_path_iterator)
-
-    plt.axes([-0.1, -0.1, 1.2, 1.2])
-    plt.pcolor([[0]], cmap='Greens')
-    mlab.test_plot3d()
-    image_rst = sg.save_figures(block, block_vars, gallery_conf)
-    assert len(plt.get_fignums()) == 0
-    assert len(image_path_iterator) == 2
-    assert '/image0.png' not in image_rst
-    assert '/image1.png' in image_rst
-    assert '/image2.png' in image_rst
-    assert '/image3.png' not in image_rst
-    assert not os.path.isfile(fname_template.format(0))
-    assert os.path.isfile(fname_template.format(1))
-    assert os.path.isfile(fname_template.format(2))
-    assert not os.path.isfile(fname_template.format(0))
-    with Image.open(fname_template.format(1)) as img:
-        pixels = np.asarray(img.convert("RGB"))
-    assert (pixels == [247, 252, 245]).all()  # plt first
-
-    # Test next-value handling, plus image_scrapers modification
-    gallery_conf.update(image_scrapers=(sg.matplotlib_scraper,))
-    mlab.test_plot3d()
-    plt.axes([-0.1, -0.1, 1.2, 1.2])
-    plt.pcolor([[0]], cmap='Reds')
-    image_rst = sg.save_figures(block, block_vars, gallery_conf)
-    assert len(plt.get_fignums()) == 0
-    assert len(image_path_iterator) == 3
-    assert '/image1.png' not in image_rst
-    assert '/image2.png' not in image_rst
-    assert '/image3.png' in image_rst
-    assert '/image4.png' not in image_rst
-    assert not os.path.isfile(fname_template.format(0))
-    for ii in range(3):
-        assert os.path.isfile(fname_template.format(ii + 1))
-    assert not os.path.isfile(fname_template.format(4))
-    with Image.open(fname_template.format(3)) as img:
-        pixels = np.asarray(img.convert("RGB"))
-    assert (pixels == [255, 245, 240]).all()
-
-    # custom finders
-    gallery_conf.update(image_scrapers=[lambda x, y, z: ''])
-    image_rst = sg.save_figures(block, block_vars, gallery_conf)
-    assert len(image_path_iterator) == 3
-
-    # degenerate
-    gallery_conf.update(image_scrapers=['foo'])
-    with pytest.raises(ValueError, match='Unknown image scraper'):
-        _complete_gallery_conf(
-            gallery_conf, gallery_conf['gallery_dir'], True, False)
-    gallery_conf.update(
-        image_scrapers=[lambda x, y, z: y['image_path_iterator'].next()])
-    with pytest.raises(RuntimeError, match='did not produce expected image'):
-        sg.save_figures(block, block_vars, gallery_conf)
-    gallery_conf.update(image_scrapers=[lambda x, y, z: 1.])
-    with pytest.raises(TypeError, match='was not a string'):
-        sg.save_figures(block, block_vars, gallery_conf)
-
-
 def test_zip_notebooks(gallery_conf):
     """Test generated zipfiles are not corrupt"""
     gallery_conf.update(examples_dir='examples')
@@ -404,42 +303,6 @@ def test_rst_example(gallery_conf):
         rst = f.read()
 
         assert "lab/tree/notebooks/plot.ipy" in rst
-
-
-def test_figure_rst():
-    """Testing rst of images"""
-    figure_list = ['sphx_glr_plot_1.png']
-    image_rst = sg.figure_rst(figure_list, '.')
-    single_image = """
-.. image:: /sphx_glr_plot_1.png
-    :class: sphx-glr-single-img
-"""
-    assert image_rst == single_image
-
-    image_rst = sg.figure_rst(figure_list + ['second.png'], '.')
-
-    image_list_rst = """
-.. rst-class:: sphx-glr-horizontal
-
-
-    *
-
-      .. image:: /sphx_glr_plot_1.png
-            :class: sphx-glr-multi-img
-
-    *
-
-      .. image:: /second.png
-            :class: sphx-glr-multi-img
-"""
-    assert image_rst == image_list_rst
-
-    # test issue #229
-    local_img = [os.path.join(os.getcwd(), 'third.png')]
-    image_rst = sg.figure_rst(local_img, '.')
-
-    single_image = sg.SINGLE_IMAGE % "third.png"
-    assert image_rst == single_image
 
 
 class TestLoggingTee:
@@ -500,6 +363,7 @@ class TestLoggingTee:
 
         monkeypatch.setattr(self.tee.output_file, 'isatty', lambda: True)
         assert self.tee.isatty()
+
 
 # TODO: test that broken thumbnail does appear when needed
 # TODO: test that examples are not executed twice
