+++ sphinx_gallery/utils.py
@@ -39,14 +39,54 @@ class _TempDir(str):
         rmtree(self._path, ignore_errors=True)
 
 
+def scale_image(in_fname, out_fname, max_width, max_height):
+    """Scales an image with the same aspect ratio centered in an
+       image box with the given max_width and max_height
+       if in_fname == out_fname the image can only be scaled down
+    """
+    # local import to avoid testing dependency on PIL:
+    try:
+        from PIL import Image
+    except ImportError:
+        import Image
+    img = Image.open(in_fname)
+    width_in, height_in = img.size
+    scale_w = max_width / float(width_in)
+    scale_h = max_height / float(height_in)
+
+    if height_in * scale_w <= max_height:
+        scale = scale_w
+    else:
+        scale = scale_h
+
+    if scale >= 1.0 and in_fname == out_fname:
+        return
+
+    width_sc = int(round(scale * width_in))
+    height_sc = int(round(scale * height_in))
+
+    # resize the image using resize; if using .thumbnail and the image is
+    # already smaller than max_width, max_height, then this won't scale up
+    # at all (maybe could be an option someday...)
+    img = img.resize((width_sc, height_sc), Image.BICUBIC)
+    # img.thumbnail((width_sc, height_sc), Image.BICUBIC)
+    # width_sc, height_sc = img.size  # necessary if using thumbnail
+
+    # insert centered
+    thumb = Image.new('RGBA', (max_width, max_height), (255, 255, 255, 255))
+    pos_insert = ((max_width - width_sc) // 2, (max_height - height_sc) // 2)
+    thumb.paste(img, pos_insert)
+
+    thumb.save(out_fname)
+
+
 def replace_py_ipynb(fname):
     """Replace .py extension in filename by .ipynb"""
-
     fname_prefix, extension = os.path.splitext(fname)
-    allowed_extension='.py'
+    allowed_extension = '.py'
     if extension != allowed_extension:
         raise ValueError(
-            "Unrecognized file extension, expected %s, got %s" % (allowed_extension, extension))
-
+            "Unrecognized file extension, expected %s, got %s"
+            % (allowed_extension, extension))
     new_extension = '.ipynb'
     return '{}{}'.format(fname_prefix, new_extension)
